Moving mpi test utility files outside cpp test folder
